Refactored the whole library.
Selectors now don't rely on ResourceDefinitions. In fact,
ResourceDefinitions is not part of NgrxJsonApiStore any more. It is
still needed for the api to obtain the 'collectionPath'.

No need to initialise the store anymore. The sole purpose of initialsation
was to inject the ResourceDefinitions but since the latter is gone, so is
the former!

Implemented denormalisation feature. Thanks to github.com/denormalize.
This is extremely important for sending data to views. Storage should be
normalized but when at the display stage the data should be denormalised.

Selectors are also much simpler now.

Filtering feature is also implemented.

Test data is starting to converge. No need to repeat the same data everywhere.