Refactor resources store. Place resources according to their type.
Resources were all stored in one big array. To query resources, complicated
methods had to be used (not that complicated!). Now, resources are now objects
where the keys are the resource types and the value is another object with
the resource id as keys.

Before:
```
resources = [
    {'type': 'Person', id: '1', name: 'Peter'},
    {'type': 'Blog', id: '1', name: 'Blog 1'}
]
```

Now:
```ts
resources = {
    'Person' : {
        '1': {type: 'Person', id: '1', name: 'Peter'}
    },
    'Blog': {
        '1': {type: 'Blog', id: '1', name: 'Blog 1'}
    }
}
```

I think this is similar to how `normalizr` works. I also think it is better in terms of
performace, to get a resource it's simply `resources[resourceType][resourceId]` rather
than complicated filtering and so on.