Refactor actions. Change actions to individual classes.
Every action is now a class. I don't fully understand why but one
can only be great if he/she stands on the shoulder of giants. Those
giants are the @ngrx guys who recommend this.

I do understand why a little, it's to do with better TypeScript tooling.

Also, the action types are now registered using a `type` function
to ensure uniqueness. It's good for large applications.