Add related objects even if they aren't included.
Previously, if an object had a relation as such:

```
let article = {
    id: '1',
    type: 'Article'
    attributes: { name: 'Article 1' },
    relationships: {
        author: {
            data: { type: 'Person', id: '1' }
        }
    }
```

Once this `Resource` is added/updated, the related object is not stored.
This causes a problem with denormalisation. So it happens many times that
`resources['Person']['1']` is undefined. This commit changes this behaviour.

Related resources are always added now.

Also, minor fixes for selectors.