Make selectors type more explicit
As a pythonista, we believe in "Explicit is better than Implicit".
Previously, the used selector depended on whether `id` or `type` are
given or not. For example:

```ts
get$({type: 'Course'})
get$({type: 'Course', id: '1'})
get$({})
```

These three gave different results because each used a different
selector internally. It is better for the user to specify which
selector he/she wants to use. For example, this helps in asynchronous
code where the `id` is not available immediately (initially undefined).
We really don't want all the results. We wait for the id to be resolved
and we'll use `get$('getOne'.....)`.

So the above examples now become:
```ts
get$('getMany', {type: 'Course'})
get$('getOne', {type: 'Course', id: '1'})
get$('getAll', {})
```