Query the store without sending requests to API (#23)
This commit added two actions that allow querying the store without sending any requests to API.

The first action `QUERY_STORE_INIT` adds the action and as a side effect the query is evaluated and sent as the payload of the other action `QUERY_STORE_SUCCESS`. A reducer will handle the latter action and save the query results to the store.

In future PRs, filtering and sorting features will be added.