Refactor selectors to allow composing
Selectors in redux are usually written in a way that allows them
to be combined. In addition, they do not depend on the actual state
(i.e. they do not use `Store` directly). This PR refactored selectors
to be like that and hence modified the NgrxJsonApiService to utilise
these selectors.