Client side filtering (#39)
* Add filtering configuration in module config

By adding filtering configuration, the path separator for filter
paths can be modified. In addition, new filtering operators can be
added or previous ones overridden. This is of course mostly theoretical,
not much has been implemented yet!

* Add client-side filtering

The queryStore selector (or query store action) will now handle
filtering params included in the query. Tests to be added in another PR.

* Add filtering support for getOne queries

* Add tests for selectors with filters

* Allow custom filtering operators

The filtering utility has a set of defined operators that one can use.
This commit allow users to add their own operators for comparing values.
The operators are part of `filteringConfig` which is itself part of
`NgrxJsonApiModuleConfig`.