Add denormalisation function with tests
Fixed the implementation for denormalisation and added some tests.

There are three types of denormalisations:
1- Resource:
  This will return a Resource object with nested Resource objects
2- PersistedResource:
  This will return a PersistedResource with nested PersistedResource
  objects (not tested!)
3- ResourceStore:
  This will return a ResourceStore with nested ResourseStore objects.
  It looks ugly: `article.resource.author.resource.name` but it may be
  useful for something like `article.resource.author.loading?`.

The default type is "Resource" which returns an object that can be used
like: `article.author.name`, `article.title`, `article.comments[0].text`

Add basic implementation for denormalise service

This implementation makes `denormalise` a selector, so after receiveing
the Observable of ResourceStore (e.g. `service.selectResourceStore`)
it can be passed to denormalise as such (`.let(denormalise())`);

Refactor denormalisation

There is no more `DenormalisationType`. If the denormalise function
is given a `Resource` it will return a denormalised Resource and if
it is given a `ResourceStore` it will return a denormalised
`ResourceStore`.

The structure of the denormalised Resource will be similar to the
normalised one with a `reference` property added to the relationships.

Example:
```ts
let obj = {
  type: 'Article',
  id: '1',
  attributes: {
    name: 'Article 1'
  },
  relationships: {
    author: {
      data: {type: 'Person', id: '1'}
    },
    comments: {
      data: [
        { type: 'Comment', id: '1'},
        { type: 'Comment', id: '2'}
      ]
    }
  }
}
```

The above Resource will be denormalised in a way that allows accessing
the author or comments resources. For example, to access the author name
: `obj.relationships.author.reference.attributes.name`. To access the
`content` of the second comment from this resource:
`obj.relationships.comments.reference[1].attributes.content`.

A similar approach is taken for denormalised ResourceStore, so the above
path will translate to:
```ts
obj.resource.relationships.comments.reference[1]
.resource.attributes.content`
```

Add tests for denormalise service and fix effects

Tests for reducers have been commented out for now because one
of the test is leaking which causes side effects in other (service)
tests.

Effects had a small bug.

Add denormaliseResource pipe

Added a denormaliseResource pipe and moved all pipes to a new file
(pipes.ts). Also added tests for the pipe.

Uncomment reducer tests

No test was leaking. It just works!

Query results selector returns ResourceStore

The `selectResults` method of `NgrxJsonApiService` uses
the `getResults$` selector. The latter used another selector
to return a Resource. This has changed now where `getResults$` now
return a ResourceStore. The reason behind this is that it's possible
to get a Resource from ResourceStore but the opposite is not true.

To obtain a Resource, a new optional argument has been added to
`service.findOne` and `service.findMany`. The third argument when set
to `true` will return a Resource rather than a ResourceStore. Its
default value is `false`.

Allow service methods to init read/patch/delete (#43)

* Allow service methods to init read/patch/delete

Previously, service methods used the "commit" method where they either
add a resource or update it or mark it for deletion. Afterwards a commit
action processes all these. This is useful in many situations. However,
sometimes it is required to perform a patch/post/delete immediately
without using commit.

This PR adds this possibility, for example, to patch a resource and send
the request to the server immediately we use:
```ts
this.service.patch(resource, true)
```

The optional boolean when set to `true` will send `API_DELETE_INIT` or
`API_UPDATE_INIT` or `API_CREATE_INIT` actions.

* Fix errors that stopped build process

Various errors stopped building the library. Mostly, TypeScript errors.
Also, `ngc` doesn't work with TS 2.1 yet so rolled back to TS 2.0.10.

* Rename commit action

Renamed all commit action to apply (e.g. API_APPLY_INIT) and also
removed the payload for ApiApplyInit. It used to have the
`storeLocation` as a payload, now it's not required so it was removed.

All this resulted in refactoring elsewhere especially in `effects.ts`
which need testing ASAP.