Refactor denormalisation
There is no more `DenormalisationType`. If the denormalise function
is given a `Resource` it will return a denormalised Resource and if
it is given a `ResourceStore` it will return a denormalised
`ResourceStore`.

The structure of the denormalised Resource will be similar to the
normalised one with a `reference` property added to the relationships.

Example:
```ts
let obj = {
  type: 'Article',
  id: '1',
  attributes: {
    name: 'Article 1'
  },
  relationships: {
    author: {
      data: {type: 'Person', id: '1'}
    },
    comments: {
      data: [
        { type: 'Comment', id: '1'},
        { type: 'Comment', id: '2'}
      ]
    }
  }
}
```

The above Resource will be denormalised in a way that allows accessing
the author or comments resources. For example, to access the author name
: `obj.relationships.author.reference.attributes.name`. To access the
`content` of the second comment from this resource:
`obj.relationships.comments.reference[1].attributes.content`.

A similar approach is taken for denormalised ResourceStore, so the above
path will translate to:
```ts
obj.resource.relationships.comments.reference[1]
.resource.attributes.content`
```