Query results selector returns ResourceStore
The `selectResults` method of `NgrxJsonApiService` uses
the `getResults$` selector. The latter used another selector
to return a Resource. This has changed now where `getResults$` now
return a ResourceStore. The reason behind this is that it's possible
to get a Resource from ResourceStore but the opposite is not true.

To obtain a Resource, a new optional argument has been added to
`service.findOne` and `service.findMany`. The third argument when set
to `true` will return a Resource rather than a ResourceStore. Its
default value is `false`.