Allow service methods to init read/patch/delete
Previously, service methods used the "commit" method where they either
add a resource or update it or mark it for deletion. Afterwards a commit
action processes all these. This is useful in many situations. However,
sometimes it is required to perform a patch/post/delete immediately
without using commit.

This PR adds this possibility, for example, to patch a resource and send
the request to the server immediately we use:
```ts
this.service.patch(resource, true)
```

The optional boolean when set to `true` will send `API_DELETE_INIT` or
`API_UPDATE_INIT` or `API_CREATE_INIT` actions.