Allow service methods to init read/patch/delete (#43)
* Allow service methods to init read/patch/delete

Previously, service methods used the "commit" method where they either
add a resource or update it or mark it for deletion. Afterwards a commit
action processes all these. This is useful in many situations. However,
sometimes it is required to perform a patch/post/delete immediately
without using commit.

This PR adds this possibility, for example, to patch a resource and send
the request to the server immediately we use:
```ts
this.service.patch(resource, true)
```

The optional boolean when set to `true` will send `API_DELETE_INIT` or
`API_UPDATE_INIT` or `API_CREATE_INIT` actions.

* Fix errors that stopped build process

Various errors stopped building the library. Mostly, TypeScript errors.
Also, `ngc` doesn't work with TS 2.1 yet so rolled back to TS 2.0.10.

* Rename commit action

Renamed all commit action to apply (e.g. API_APPLY_INIT) and also
removed the payload for ApiApplyInit. It used to have the
`storeLocation` as a payload, now it's not required so it was removed.

All this resulted in refactoring elsewhere especially in `effects.ts`
which need testing ASAP.