Remove QueryHandle and implement new logic for unsubscribing
When unsubscribing from the Observable streaming the query results,
we want to also remove the query itself. This was done by extending
the AnonymousSubscription and calling a custom function when running
`unsubscribe`. Rather than doing this, we use the `rxjs` `finally`
operator to run this "teardown logic".

Tests have also been added for `service.findOne` and
`service.findMany`.