Add denormaliseOne and denormaliseMany functions
Due to TypeScript type check, I couldn't figure out how to do
a `denormalise` function that would take EITHER a StoreResource or
an Array of StoreResource (StoreResource[]) and then outputs a
DenormalisedStoreResource or an Array of DenormalisedStoreResource.

So for the time being until I figure out how to do it, life will go
on with two denormalise functions. One for a single StoreResource
and the other for multiple StoreResource. Use wisely!