Always change the operation status
Previously, when trying to create or update a resource which is
already in the store (exactly as it is) the reducers responsible
will ignore the new data and just return the original state.
This has a side effect of ignoring to change isUpdating or
isCreating which is essential when using the library.

The status update is mandatory regardless if the data is already in
the store or not.